# docs/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Nova HTTP API
  version: 0.2.0-beta.1+ad629d7
  summary: Federated content-addressed blob storage with image transforms.
  description: |
    The public-facing HTTP API surface of a Nova deployment.

    This document is the source of truth for handler codegen
    (`oapi-codegen` for Go), the drag-and-drop upload widget
    (`openapi-typescript` for the TS client), and downstream language
    SDKs. CI fails the build if handler signatures and this document
    drift. See `docs/specs/PRODUCT_MODULE_INTERFACE.md` for how product
    layers like `nova-image` plug into the storage core's URL surface.

    Two distinct namespaces:

      - **Content** routes (`/blob/{cid}`, `/i/{cid}/...`) are
        unversioned, immutable, and aggressively CDN-cacheable. They
        serve bytes; they do not change.
      - **Management** routes (`/api/v1/...`) are versioned and require
        authentication unless the operator runs in `auth: anonymous`
        mode. Breaking changes go through a major version bump.

    Federation routes (`/fed/v1/...`) are reachable only over the
    operator's mesh VPN and use mutual TLS via Nebula certificates.
    They are documented separately in
    `docs/specs/FEDERATION_PROTOCOL.md`; they do not appear here
    because no public consumer should reach them.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://{host}
    description: An operator's public Nova endpoint.
    variables:
...
security:
  - {}                # default: public read for content routes
  - signedUrl: []     # private content via HMAC-signed URLs
  - bearerAuth: []    # management endpoints
...
paths:
...
  /blob/{cid}:
    parameters:
      - $ref: '#/components/parameters/CidParam'
    get:
      tags: [content-blob]
      operationId: getBlob
      summary: Retrieve the bytes of a blob by CID.
      description: |
        Returns the bytes verbatim with the stored `Content-Type`. For
        encrypted collections the bytes are plaintext (the gateway
        decrypts on the way out). For collections the operator marked
        `public_archival: true`, the bytes are stored unencrypted and
        the gateway streams them directly.

        **Cache-Control varies by collection visibility and blob state**
        (v2 stratified policy):

        | collection | state | request | Cache-Control |
        |---|---|---|---|
        | public | active | anonymous | `public, max-age=31536000, immutable` |
        | public-archival | active | anonymous | `public, max-age=31536000, immutable` |
        | unlisted | active | anonymous (CID is the capability) | `private, max-age=300, must-revalidate` |
        | private | active | signed URL / bearer | `private, max-age=300, must-revalidate` |
        | any | soft_deleted/quarantined/tombstoned | any | `no-store` |

        **Range requests are supported only on public-archival
        (plaintext) blobs.** A Range request against an encrypted
        blob returns `416 Range Not Satisfiable`. Streaming AEAD
        is a Phase 2 deliverable; until then, single-shot decryption
        cannot serve byte ranges.
      parameters:
        - $ref: '#/components/parameters/RangeHeader'
      responses:
        '200':
          $ref: '#/components/responses/BlobBytes'
        '206':
          $ref: '#/components/responses/PartialBlobBytes'
        '401':
          $ref: '#/components/responses/SignedUrlRequired'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Tombstoned'
        '416':
          $ref: '#/components/responses/RangeNotSatisfiable'
        '451':
          $ref: '#/components/responses/Quarantined'
    head:
      tags: [content-blob]
      operationId: headBlob
      summary: Probe a blob's existence and metadata via headers.
      responses:
        '200':
          description: Blob exists; metadata returned in headers.
          headers:
            Content-Type: { schema: { type: string } }
            Content-Length: { schema: { type: integer, format: int64 } }
            ETag: { schema: { type: string } }
            X-Nova-Cid: { schema: { type: string } }
        '404': { $ref: '#/components/responses/NotFound' }
        '410': { $ref: '#/components/responses/Tombstoned' }
...
  /api/v1/uploads:
    post:
      tags: [upload]
      operationId: createUpload
      summary: Create a resumable upload (tus).
      description: |
        Implements the [tus.io](https://tus.io) resumable upload
        protocol v1.0.0. The response `Location` header points at the
        upload-specific endpoint that subsequent `PATCH`, `HEAD`, and
        `DELETE` requests target.

        After the final `PATCH` completes, call `POST .../finalize` to
        commit the upload, compute the CID, and write the metadata
        row.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/offset+octet-stream:
            example: |
              # tus creation-with-upload: the body is the first chunk of raw bytes.
              # Length and metadata travel in headers, not in the body:
              #   Upload-Length: 20481
              #   Upload-Metadata: filename cGhvdG8uanBn,product aW1hZ2U=
              #   Content-Type: application/offset+octet-stream
              <20481 bytes of image/jpeg>
            schema: { type: string, format: binary }
      parameters:
        - in: header
          name: Tus-Resumable
          required: true
          schema: { type: string, enum: ['1.0.0'] }
        - in: header
          name: Upload-Length
          required: true
          schema: { type: integer, format: int64, minimum: 0 }
        - in: header
          name: Upload-Metadata
          required: false
          description: |
            tus metadata field. Nova recognizes:

              - `filename` — original filename (informational)
              - `mime_type` — declared content-type (validated server-side)
              - `product` — `image`, `video`, `audio`, `archive`,
                `document`, `raw` — selects the validation pipeline.
              - `collection_id` — UUID of the destination collection.

            (v2: `expected_cid` was removed. Random-nonce envelope
            encryption means the client cannot compute the envelope
            CID, so the field had no meaningful semantics. Plaintext-
            hash dedup, if reintroduced later, will use a different
            field name to avoid confusion.)
          schema: { type: string }
      responses:
        '201':
          description: Upload created.
          headers:
            Location:
              required: true
              schema: { type: string, format: uri }
              description: The upload-specific URL for subsequent calls.
            Tus-Resumable: { schema: { type: string } }
            Upload-Offset: { schema: { type: integer, format: int64 } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413':
          description: >-
            The declared or actual body length exceeds this operator's
            `max_upload_size`. The limit is enforced before the bytes
            are accepted, so nothing was stored and no upload session
            remains to resume. Split the object or ask the operator to
            raise the ceiling.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/uploads/{id}:
    parameters:
      - $ref: '#/components/parameters/UploadIdParam'
    head:
      tags: [upload]
      operationId: getUploadOffset
      summary: Probe the current offset of an in-progress upload.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Offset and metadata returned in headers.
          headers:
            Upload-Offset: { schema: { type: integer, format: int64 } }
            Upload-Length: { schema: { type: integer, format: int64 } }
            Tus-Resumable: { schema: { type: string } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [upload]
      operationId: appendUploadChunk
      summary: Append a chunk to an in-progress upload (tus PATCH).
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: Upload-Offset
          required: true
          schema: { type: integer, format: int64 }
        - in: header
          name: Tus-Resumable
          required: true
          schema: { type: string, enum: ['1.0.0'] }
        - in: header
          name: Content-Type
          required: true
          schema: { type: string, enum: ['application/offset+octet-stream'] }
      requestBody:
        required: true
        content:
          application/offset+octet-stream:
            example: |
              # Each PATCH carries the next contiguous run of bytes. Upload-Offset must
              # equal the server's current offset or the request is rejected with 409:
              #   Upload-Offset: 16384
              #   Content-Type: application/offset+octet-stream
              <4097 bytes continuing the object>
            schema: { type: string, format: binary }
      responses:
        '204':
          description: Chunk accepted.
          headers:
            Upload-Offset: { schema: { type: integer, format: int64 } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: >-
            `Upload-Offset` does not match the offset the coordinator
            holds for this session, so the chunk would leave a gap or
            overwrite bytes already committed. Issue `HEAD
            /api/v1/uploads/{id}` to read the authoritative offset and
            resume from exactly there. This is the normal tus recovery
            path after a dropped connection, not an error condition in
            itself.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [upload]
      operationId: abandonUpload
      summary: Abandon an in-progress upload.
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Upload deleted. }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/blobs:
    post:
      tags: [upload]
      operationId: uploadBlobMultipart
      summary: Upload a generic blob via multipart/form-data.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            example:
              file: <binary contents of report.pdf>
              product: document
              collection_id: 31818242-97c3-4cdf-8025-44ffcb3d0242
              alt_text: Q3 availability report
              caption: Generated from the coordinator's own metrics.
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                product:
                  type: string
                  enum: [image, video, audio, archive, document, raw]
                  default: raw
                collection_id:
                  type: string
                  format: uuid
                alt_text: { type: string }
                caption: { type: string }
      responses:
        '201':
          description: Blob created.
          content:
            application/json:
              example:
                cid: bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi
                byte_size: 204815
                mime_type: image/jpeg
                product: image
                urls:
                  original: https://nova.thebugreport.dev/i/bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi
                  json: https://nova.thebugreport.dev/i/bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi.json
                  presets:
                    thumb: https://nova.thebugreport.dev/i/bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi/p/thumb.webp
                    medium: https://nova.thebugreport.dev/i/bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi/p/medium.webp
              schema: { $ref: '#/components/schemas/UploadResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413':
          description: >-
            The declared or actual body length exceeds this operator's
            `max_upload_size`. The limit is enforced before the bytes
            are accepted, so nothing was stored and no upload session
            remains to resume. Split the object or ask the operator to
            raise the ceiling.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/images:
    post:
      tags: [upload]
      operationId: uploadImageMultipart
      summary: Upload an image with image-specific validation.
      description: |
        Identical to `POST /api/v1/blobs` with `product=image`, but
        rejects non-image MIME types early and checks the CID against
        the operator-curated CID blocklist so the caller learns
        immediately if the upload was rejected on moderation grounds.
        The blocklist is a direct indexed lookup; no perceptual scan is
        performed at upload time (Phase-3 deferral).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            example:
              file: <binary contents of photo.jpg>
              collection_id: 31818242-97c3-4cdf-8025-44ffcb3d0242
              alt_text: A patch panel with sixteen cables terminated
              caption: The rack as it stood in July.
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                collection_id:
                  type: string
                  format: uuid
                alt_text: { type: string }
                caption: { type: string }
      responses:
        '201':
          description: Image created.
          content:
            application/json:
              example:
                cid: bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi
                byte_size: 204815
                mime_type: image/jpeg
                product: image
                urls:
                  original: https://nova.thebugreport.dev/i/bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi
                  json: https://nova.thebugreport.dev/i/bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi.json
                  presets:
                    thumb: https://nova.thebugreport.dev/i/bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi/p/thumb.webp
                    medium: https://nova.thebugreport.dev/i/bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi/p/medium.webp
              schema: { $ref: '#/components/schemas/UploadResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '415':
          description: >-
            The uploaded file's detected media type is not one this
            coordinator accepts as an image. Detection is by content
            sniffing rather than by the supplied filename or `Content-
            Type`, so renaming the file will not change the outcome.
            Upload it through `POST /api/v1/blobs` if you want it stored
            as a non-image product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '451':
          description: >-
            The content hashes to a CID on this operator's blocklist and
            was refused at ingest. Because the address is derived from
            the bytes, an identical file will always be refused; nothing
            was stored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # =========================================================================
  # Metadata management
  # =========================================================================
...
  /api/v1/images/{cid}:
    parameters:
      - $ref: '#/components/parameters/CidParam'
    get:
      tags: [metadata]
      operationId: getImageMeta
      summary: Authenticated metadata read for an image.
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Image metadata.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Image' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [metadata]
      operationId: updateImageMeta
      summary: Update mutable owner-controlled fields on an image.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            example:
              alt_text: A patch panel with sixteen cables terminated
              caption: The rack as it stood in July.
            schema: { $ref: '#/components/schemas/ImageUpdate' }
      responses:
        '200':
          description: Image metadata updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Image' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [metadata]
      operationId: deleteImage
      summary: Soft-delete an image (alias for `DELETE /api/v1/blobs/{cid}`).
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Deletion accepted. }
        '404': { $ref: '#/components/responses/NotFound' }

  # =========================================================================
  # Collections
  # =========================================================================
...
  /api/v1/auth/refresh:
    post:
      tags: [admin]
      operationId: authRefresh
      summary: Rotate a refresh token for a new pair (single-use; reuse revokes the family).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              refresh_token: v1.MEQCIB3q...c2VjcmV0.9f2b1a7c
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: New token pair.
          content:
            application/json:
              example:
                access_token: eyJhbGciOiJFZERTQSIsImtpZCI6InNpZy0yMDI2LTA3In0...
                refresh_token: v1.MEQCIFp7...c2VjcmV0.31d8c04a
                token_type: bearer
                expires_in: 900
                kid: sig-2026-07
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401':
          description: >-
            The refresh token was rejected. It was expired, already
            revoked, or — the case worth knowing about — presented a
            second time. Refresh tokens are single-use, and replay is
            treated as evidence of theft: the whole token family is
            revoked, which logs out every session descended from that
            login. Authenticate again with credentials.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: >-
            This coordinator is configured to delegate authentication to
            an external OIDC provider, so its own token issuer is
            switched off and the route does not exist. Obtain a token
            from the configured provider instead; `GET
            /api/v1/auth/config` reports which mode is active.
  /api/v1/auth/logout:
    post:
      tags: [admin]
      operationId: authLogout
      summary: Revoke a refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            example:
              refresh_token: v1.MEQCIB3q...c2VjcmV0.9f2b1a7c
            schema: { $ref: '#/components/schemas/LogoutRequest' }
      responses:
        '204': { description: Logged out (idempotent). }
        '404':
          description: >-
            This coordinator is configured to delegate authentication to
            an external OIDC provider, so its own token issuer is
            switched off and the route does not exist. Obtain a token
            from the configured provider instead; `GET
            /api/v1/auth/config` reports which mode is active.
...
components:
  securitySchemes:

    signedUrl:
      type: apiKey
      in: query
      name: sig
      description: |
        HMAC-SHA256 signature, base64url-encoded. Always accompanied by
        `exp`, `aud`, and `kid` query parameters. See
        `docs/specs/SIGNED_URL_FORMAT.md`.

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token. The **Phase-1 default** is Nova's built-in local
        OIDC-shaped issuer (EdDSA/Ed25519; see `/api/v1/auth/*`). Operators may
        instead point `auth.issuer_url` at an external OIDC provider (Authelia or
        any compatible IdP), in which case the coordinator verifies external
        tokens via the IdP's JWKS and the local-issuer endpoints return 404
        `external_oidc_active`. Both verify concurrently during migration. The
        coordinator is a resource server: it validates bearer tokens but does not
        mediate the interactive login flow (the SPA/CLI drive PKCE, discovering
        the IdP via `/api/v1/auth/config`).

        Management `/api/v1/*` writes and admin routes require a bearer with a
        sufficient role; public content reads (`/blob/*`, `/i/*`) do not.

        v2 note: the original `auth: anonymous` mode for management endpoints was
        removed. It is available only in dev builds (build tag `nova_dev`) and
        only when the coordinator binds to loopback. Production builds refuse to
        start with anonymous management auth regardless of operator config.
...